Update README.md
Created a Browser based Tic-Tac-Toe Game in which Two players can play a game on the Local Browser using HTML, CSS and JavaScript to build this game.
Tic-Tac-Toe is a two-player game in which the players fill up nine empty rectangles in a table with either an X or an O when it is their turn. Once someone succeeds to line up their sign vertically or horizontally without an interruption, that player wins.